refactor(queue): Move completion tracking behind runtime SPI
Move the remaining QueueToadlet completion-tracking slice into a\nlegacy runtime adapter behind RuntimePorts#queueCompletion().\n\nKeep QueueToadlet focused on HTTP responsibilities, move\ncompletion persistence/replay and alert registration into\nLegacyQueueCompletionPort, and refresh the architecture docs\nfor the expanded runtime SPI surface.
